Convert Room Member and User to TS
